+++ addons/confirma/src/classes/TestingMethod.cs
@@ -15,7 +15,7 @@ public class TestingMethod
     public MethodInfo Method { get; }
     public IEnumerable<TestCase> TestCases { get; }
     public string Name { get; }
-    public TestMethodResult Result { get; private set; }
+    public TestMethodResult Result { get; }
 
     public TestingMethod(MethodInfo method)
     {
@@ -31,8 +31,8 @@ public class TestingMethod
         {
             for (ushort i = 0; i <= test.Repeat; i++)
             {
-                var attr = test.Method.GetCustomAttribute<IgnoreAttribute>();
-                if (attr is not null && attr.IsIgnored())
+                IgnoreAttribute? attr = test.Method.GetCustomAttribute<IgnoreAttribute>();
+                if (attr?.IsIgnored() == true)
                 {
                     Result.TestsIgnored++;
 
@@ -64,43 +64,42 @@ public class TestingMethod
     private IEnumerable<TestCase> DiscoverTestCases()
     {
         List<TestCase> cases = new();
-        var discovered = TestDiscovery.GetTestCasesFromMethod(Method).GetEnumerator();
+        using IEnumerator<System.Attribute> discovered =
+            TestDiscovery.GetTestCasesFromMethod(Method).GetEnumerator();
 
         while (discovered.MoveNext())
         {
-            if (discovered.Current is TestCaseAttribute testCase)
+            switch (discovered.Current)
             {
-                cases.Add(new(Method, testCase.Parameters, 0));
-                continue;
-            }
-
-            // I rely on the order in which the attributes are defined
-            // to determine which TestCase attributes should be assigned values
-            // from the Repeat attributes.
-            if (discovered.Current is RepeatAttribute repeat)
-            {
-                if (!discovered.MoveNext())
-                {
+                case TestCaseAttribute testCase:
+                    cases.Add(new(Method, testCase.Parameters, 0));
+                    continue;
+                // I rely on the order in which the attributes are defined
+                // to determine which TestCase attributes should be assigned values
+                // from the Repeat attributes.
+                case RepeatAttribute when !discovered.MoveNext():
                     Log.PrintWarning(
                         $"The Repeat attribute for the \"{Method.Name}\" method will be ignored " +
                         "because it does not have the TestCase attribute after it.\n"
                     );
                     Result.Warnings++;
                     continue;
-                }
-
-                if (discovered.Current is RepeatAttribute)
-                {
+                case RepeatAttribute when discovered.Current is RepeatAttribute:
                     Log.PrintWarning(
                         $"The Repeat attributes for the \"{Method.Name}\" cannot occur in succession.\n"
                     );
                     Result.Warnings++;
                     continue;
-                }
+                case RepeatAttribute repeat:
+                    {
+                        if (discovered.Current is not TestCaseAttribute tc)
+                        {
+                            continue;
+                        }
 
-                if (discovered.Current is not TestCaseAttribute tc) continue;
-
-                cases.Add(new(Method, tc.Parameters, repeat.Repeat));
+                        cases.Add(new(Method, tc.Parameters, repeat.Repeat));
+                        break;
+                    }
             }
         }