+++ addons/confirma/src/scenes/confirma_autoload/ConfirmaAutoload.cs
@@ -1,3 +1,4 @@
+using System;
 using Confirma.Classes;
 using Confirma.Helpers;
 using Confirma.Types;
@@ -14,7 +15,10 @@ public partial class ConfirmaAutoload : Node
     {
         CheckArguments();
 
-        if (!Props.RunTests) return;
+        if (!Props.RunTests)
+        {
+            return;
+        }
 
         SetupGlobals();
         ChangeScene();
@@ -30,11 +34,14 @@ public partial class ConfirmaAutoload : Node
     {
         string[] args = OS.GetCmdlineUserArgs();
 
-        if (DisplayServer.GetName() == "headless") Props.IsHeadless = true;
-
-        foreach (var arg in args)
+        if (DisplayServer.GetName() == "headless")
         {
-            if (!Props.RunTests && arg.StartsWith("--confirma-run"))
+            Props.IsHeadless = true;
+        }
+
+        foreach (string arg in args)
+        {
+            if (!Props.RunTests && arg.StartsWith("--confirma-run", StringComparison.InvariantCulture))
             {
                 Props.RunTests = true;
 
@@ -44,12 +51,14 @@ public partial class ConfirmaAutoload : Node
 
                 continue;
             }
-
-            if (Props.RunTests && !Props.ClassName.Equals(string.Empty) && arg.StartsWith("--confirma-method"))
+            else if (Props.RunTests
+                && !Props.ClassName.Equals(string.Empty, StringComparison.Ordinal)
+                && arg.StartsWith("--confirma-method", StringComparison.InvariantCulture)
+            )
             {
                 Props.MethodName = arg.Find('=') == -1
-                    ? string.Empty
-                    : arg.Split('=')[1];
+                                    ? string.Empty
+                                    : arg.Split('=')[1];
 
                 continue;
             }
@@ -75,15 +84,17 @@ public partial class ConfirmaAutoload : Node
             if (!Props.DisableParallelization && arg == "--confirma-sequential")
             {
                 Props.DisableParallelization = true;
-                continue;
             }
         }
     }
 
     private void ChangeScene()
     {
-        GetTree().CallDeferred("change_scene_to_file", "uid://cq76c14wl2ti3");
+        _ = GetTree().CallDeferred("change_scene_to_file", "uid://cq76c14wl2ti3");
 
-        if (Props.QuitAfterTests) GetTree().Quit();
+        if (Props.QuitAfterTests)
+        {
+            GetTree().Quit();
+        }
     }
 }