+++ CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 - '--confirma-method' argument allowing to run a single method.
 
+### Changed
+
+- Confirma uses invariant culture when working with strings & numbers.
+- TestDiscovery class is now static.
+- Renamed protected '_autoload' & '_output' fields to 'Autoload' & 'Output'.
+
 ## [0.6.1-beta 2024-07-08]
 
 ### Added